Minor Error fixes
Clear base arduino code now report battery power as float.
Touch typing report code, and file names corrected for linux.

+++ ClearBaseRobot/Battary.ino
@@ -1,4 +1,4 @@
-long Battery(){
+float Battery(){
   int sensorValue = analogRead(A0);
 
   // Convert the analog reading (which goes from 0 - 1023) to a voltage (0 - 5V):
